requirements.txt
# Core Framework
fastapi==0.110.0
uvicorn[standard]==0.27.1
python-multipart==0.0.9

# Database
sqlalchemy[asyncio]==2.0.28
asyncpg==0.29.0
alembic==1.13.1
geoalchemy2==0.14.6

# Redis
redis[hiredis]==5.0.2
aioredis==2.0.1

# Auth & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.27.0
authlib==1.3.0

# Validation & Settings
pydantic==2.6.3
pydantic-settings==2.2.1
email-validator==2.1.1

# Async utils
anyio==4.3.0
aiofiles==23.2.1

# Elasticsearch
elasticsearch[async]==8.12.1

# Task Queue
celery[redis]==5.3.6
flower==2.0.1
aiokafka==0.10.0

# Payment
razorpay==1.4.1

# Notifications
firebase-admin==6.4.0
twilio==8.13.0
resend==0.8.0

# Observability & Monitoring
opentelemetry-api==1.23.0
opentelemetry-sdk==1.23.0
opentelemetry-instrumentation-fastapi==0.44b0
opentelemetry-exporter-otlp==1.23.0
prometheus-fastapi-instrumentator==6.1.0
prometheus-client==0.19.0

# Resilience & Circuit Breaking
pybreaker==1.4.0
tenacity==8.2.3

# Utilities
python-dateutil==2.9.0
phonenumbers==8.13.32
pillow==10.2.0
python-slugify==8.0.4

# Testing
pytest==8.1.0
pytest-asyncio==0.23.5
pytest-cov==4.1.0
factory-boy==3.3.0
faker==24.3.0

itsdangerous==2.2.0
